1.增加pwm模式14 2.修改title时间不能大于256的问题 3.修复报警时没有判断开关是否开启的问题 4.修复falldown enable错误，enable应为1bit

+++ app/config/dev_config.get.h
@@ -120,7 +120,8 @@ const alert_motion_t * dev_config_get_alert_motion(void);
 
  bool dev_config_get_alert_tilt_enable(void);
  uint8_t dev_config_get_alert_tilt_angle(void);
- uint8_t dev_config_get_alert_tilt_threshold(void);
+// uint8_t dev_config_get_alert_tilt_threshold(void);
+uint16_t dev_config_get_alert_tilt_threshold(void);
  bool dev_config_get_alert_falldown_enable(void);
  bool dev_config_get_alert_falldown_call(void);
  uint8_t dev_config_get_alert_falldown_level(void);
